+++ packages/databyss-ui/theming/views.js
@@ -39,7 +39,6 @@ const borderVariants = {
   formField: border(2, 'border.1'),
   formError: border(2, 'red.0'),
   activeFormField: border(2, 'primary.0'),
-  modal: { borderRadius: pxUnits(5) },
   round: { borderRadius: '50%' },
 }
